cerbero: Only remove failed downloads if they exist
Otherwise we just mask the actual error with a new "file not found"
error.

+++ cerbero/utils/shell.py
@@ -246,7 +246,8 @@ def download(url, destination=None, recursive=False, check_cert=True, overwrite=
         try:
             call(cmd, path)
         except FatalError, e:
-            os.remove(destination)
+            if os.path.exists(destination):
+                os.remove(destination)
             raise e
